requirements.txt
python-telegram-bot==21.0.1
mysql-connector-python==8.3.0
